10476 file(1) could be smatch clean Reviewed by: Toomas Soome <tsoome@me.com> Reviewed by: Hans Rosenfeld <rosenfeld@grumpf.hope-2000.org> Reviewed by: Robert Mustacchi <rm@joyent.com> Approved by: Dan McDonald <danmcd@joyent.com>

+++ usr/src/cmd/file/elf_read.c
@@ -510,7 +510,7 @@ process_shdr(Elf_Info *EI)
 			Elf_Dyn dyn;
 			FILE_ELF_SIZE_T dsize;
 			FILE_ELF_OFF_T doff;
-			int dynn;
+			uint64_t dynn;
 
 			doff = shdr->sh_offset;
 			dsize = sizeof (Elf_Dyn);
